+++ src/runtime/src/core/io.rs
@@ -1,10 +1,10 @@
 use {
-    crate::{runtime_error, ExternalData, ExternalValue, MetaMap, Value, ValueMap},
+    crate::{runtime_error, ExternalData, ExternalValue, MetaMap, RuntimeError, Value, ValueMap},
     lazy_static::lazy_static,
     parking_lot::RwLock,
     std::{
         fmt, fs,
-        io::{Read, Seek, SeekFrom, Write},
+        io::{self, Read, Seek, SeekFrom, Write},
         path::{Path, PathBuf},
         sync::Arc,
     },
@@ -20,7 +20,7 @@ pub fn make_module() -> ValueMap {
             [Str(path)] => {
                 let path = Path::new(path.as_str());
                 match fs::File::create(&path) {
-                    Ok(file) => Ok(File::make_external_value(file, path, false)),
+                    Ok(file) => Ok(File::with_file(file, path, false)),
                     Err(e) => {
                         return runtime_error!("io.create: Error while creating file: {}", e);
                     }
@@ -44,7 +44,7 @@ pub fn make_module() -> ValueMap {
             [Str(path)] => {
                 let path = Path::new(path.as_str());
                 match fs::File::open(&path) {
-                    Ok(file) => Ok(File::make_external_value(file, path, false)),
+                    Ok(file) => Ok(File::with_file(file, path, false)),
                     Err(e) => {
                         return runtime_error!("io.open: Error while opening path: {}", e);
                     }
@@ -87,6 +87,10 @@ pub fn make_module() -> ValueMap {
         }
     });
 
+    result.add_fn("stderr", |_, _| Ok(File::stderr()));
+    result.add_fn("stdin", |_, _| Ok(File::stdin()));
+    result.add_fn("stdout", |_, _| Ok(File::stdout()));
+
     result.add_fn("temp_dir", {
         |_, _| Ok(Str(std::env::temp_dir().to_string_lossy().as_ref().into()))
     });
@@ -100,24 +104,12 @@ lazy_static! {
 
         let mut meta = MetaMap::with_type_name("File");
 
-        meta.add_named_instance_fn("path", |file: &File, _, _| {
-            Ok(Str(file.path.to_string_lossy().as_ref().into()))
-        });
+        meta.add_named_instance_fn("path", |file: &File, _, _| Ok(Str(file.path().into())));
 
         meta.add_named_instance_fn_mut("read_to_string", |file: &mut File, _, _| {
-            match file.file.seek(SeekFrom::Start(0)) {
-                Ok(_) => {
-                    let mut buffer = String::new();
-                    match file.file.read_to_string(&mut buffer) {
-                        Ok(_) => Ok(Str(buffer.into())),
-                        Err(e) => {
-                            runtime_error!("File.read_to_string: Error while reading data: {}", e,)
-                        }
-                    }
-                }
-                Err(e) => {
-                    runtime_error!("File.read_to_string: Error while seeking in file: {}", e)
-                }
+            match file.read_to_string() {
+                Ok(result) => Ok(result.into()),
+                Err(e) => Err(e.with_prefix("File.read_to_string")),
             }
         });
 
@@ -126,11 +118,9 @@ lazy_static! {
                 if *n < 0.0 {
                     return runtime_error!("File.seek: Negative seek positions not allowed");
                 }
-                match file.file.seek(SeekFrom::Start(n.into())) {
+                match file.seek(n.into()) {
                     Ok(_) => Ok(Value::Empty),
-                    Err(e) => {
-                        runtime_error!("File.seek: Error while seeking in file: {}", e)
-                    }
+                    Err(e) => Err(e.with_prefix("File.seek")),
                 }
             }
             [unexpected] => runtime_error!(
@@ -141,16 +131,10 @@ lazy_static! {
         });
 
         meta.add_named_instance_fn_mut("write", |file: &mut File, _, args| match args {
-            [value] => {
-                let data = format!("{}", value);
-
-                match file.file.write(data.as_bytes()) {
-                    Ok(_) => Ok(Value::Empty),
-                    Err(e) => {
-                        runtime_error!("File.write: Error while writing to file: {}", e)
-                    }
-                }
-            }
+            [value] => match file.write(value.to_string().as_bytes()) {
+                Ok(_) => Ok(Value::Empty),
+                Err(e) => Err(e.with_prefix("File.write")),
+            },
             _ => runtime_error!("File.write: Expected single value to write as argument"),
         });
 
@@ -162,9 +146,9 @@ lazy_static! {
                     return runtime_error!("File.write_line: Expected single value as argument");
                 }
             };
-            match file.file.write(line.as_bytes()) {
+            match file.write(line.as_bytes()) {
                 Ok(_) => Ok(Value::Empty),
-                Err(e) => runtime_error!("File.write_line: Error while writing to file: {}", e),
+                Err(e) => Err(e.with_prefix("File.write_line")),
             }
         });
 
@@ -173,31 +157,80 @@ lazy_static! {
 }
 
 #[derive(Debug)]
-pub struct File {
-    pub file: fs::File,
-    pub path: PathBuf,
-    /// When set to true the file will be removed when Dropped
-    pub temporary: bool,
+pub enum File {
+    System(SystemFile),
+    Stderr(io::Stderr),
+    Stdin(io::Stdin),
+    Stdout(io::Stdout),
 }
 
 impl File {
-    pub fn make_external_value(file: fs::File, path: &Path, temporary: bool) -> Value {
+    pub fn with_file(file: fs::File, path: &Path, temporary: bool) -> Value {
         let result = ExternalValue::with_shared_meta_map(
-            File {
+            File::System(SystemFile {
                 file,
                 path: path.to_path_buf(),
                 temporary,
-            },
+            }),
             FILE_META.clone(),
         );
         Value::ExternalValue(result)
     }
-}
 
-impl Drop for File {
-    fn drop(&mut self) {
-        if self.temporary {
-            let _ = fs::remove_file(&self.path).is_ok();
+    pub fn stderr() -> Value {
+        let stderr = File::Stderr(io::stderr());
+        let result = ExternalValue::with_shared_meta_map(stderr, FILE_META.clone());
+        Value::ExternalValue(result)
+    }
+
+    pub fn stdin() -> Value {
+        let stdin = File::Stdin(io::stdin());
+        let result = ExternalValue::with_shared_meta_map(stdin, FILE_META.clone());
+        Value::ExternalValue(result)
+    }
+
+    pub fn stdout() -> Value {
+        let stdout = File::Stdout(io::stdout());
+        let result = ExternalValue::with_shared_meta_map(stdout, FILE_META.clone());
+        Value::ExternalValue(result)
+    }
+
+    pub fn path(&self) -> String {
+        match self {
+            File::System(file) => file.path.to_string_lossy().as_ref().into(),
+            File::Stderr(_) => "_stderr_".into(),
+            File::Stdin(_) => "_stdin_".into(),
+            File::Stdout(_) => "_stdout_".into(),
+        }
+    }
+
+    pub fn read_to_string(&mut self) -> Result<String, RuntimeError> {
+        match self {
+            File::System(file) => file.read_to_string(),
+            File::Stdin(stdin) => {
+                let mut result = String::new();
+                match stdin.read_to_string(&mut result) {
+                    Ok(_) => Ok(result),
+                    Err(e) => Err(e.to_string().into()),
+                }
+            }
+            _ => runtime_error!("seek unsupported for this file type"),
+        }
+    }
+
+    pub fn seek(&mut self, position: u64) -> Result<(), RuntimeError> {
+        match self {
+            File::System(file) => file.seek(position),
+            _ => runtime_error!("seek unsupported for this file type"),
+        }
+    }
+
+    pub fn write(&mut self, bytes: &[u8]) -> Result<(), RuntimeError> {
+        match self {
+            File::System(file) => file.file.write_all(bytes).map_err(|e| e.to_string().into()),
+            File::Stderr(stderr) => stderr.write_all(bytes).map_err(|e| e.to_string().into()),
+            File::Stdout(stdout) => stdout.write_all(bytes).map_err(|e| e.to_string().into()),
+            _ => runtime_error!("seek unsupported for this file type"),
         }
     }
 }
@@ -210,6 +243,44 @@ impl ExternalData for File {
 
 impl fmt::Display for File {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "File({})", self.path.to_string_lossy())
+        write!(f, "File({})", self.path())
+    }
+}
+
+#[derive(Debug)]
+pub struct SystemFile {
+    pub file: fs::File,
+    pub path: PathBuf,
+    /// When set to true, the file will be removed when Dropped
+    pub temporary: bool,
+}
+
+impl SystemFile {
+    pub fn read_to_string(&mut self) -> Result<String, RuntimeError> {
+        match self.file.seek(SeekFrom::Start(0)) {
+            Ok(_) => {
+                let mut buffer = String::new();
+                match self.file.read_to_string(&mut buffer) {
+                    Ok(_) => Ok(buffer),
+                    Err(e) => runtime_error!("Error while reading data: {}", e),
+                }
+            }
+            Err(e) => runtime_error!("Error while seeking in file: {}", e),
+        }
+    }
+
+    pub fn seek(&mut self, position: u64) -> Result<(), RuntimeError> {
+        match self.file.seek(SeekFrom::Start(position)) {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e.to_string().into()),
+        }
+    }
+}
+
+impl Drop for SystemFile {
+    fn drop(&mut self) {
+        if self.temporary {
+            let _ = fs::remove_file(&self.path).is_ok();
+        }
     }
 }